bhv model to wk
cos sim > 0.9999 (for txt comparison)

+++ tools/data/yolo2yolo.py
@@ -28,6 +28,8 @@ def process(args):
                 path_lb_i = path_lb_i.replace('.jpg', '.txt')
                 path_lb_o = path_img_o.replace('images', 'labels')
                 path_lb_o = path_lb_o.replace('.jpg', '.txt')
+                if not os.path.exists(path_lb_i):
+                    continue
                 with open(path_lb_i, 'r') as flbi, open(path_lb_o, 'w') as flbo:
                     lines_lb_i = flbi.readlines()
                     for line_lb_i in lines_lb_i:
@@ -54,8 +56,8 @@ def new_dirs(args):
 
 def parse_args():
     parser = argparse.ArgumentParser()
-    parser.add_argument('--dir_in_root', default='/media/sdb/data/custom_head', type=str)
-    parser.add_argument('--dir_out_root', default='/media/sdb/data/custom_head_cut', type=str)
+    parser.add_argument('--dir_in_root', default='/media/sdb/data/custom_behavior', type=str)
+    parser.add_argument('--dir_out_root', default='/media/sdb/data/custom_behavior_cut', type=str)
     parser.add_argument('--subsets', nargs='*', default=['train', 'val'], type=str)
     return parser.parse_args()
